docs: Proofread for correct use of "```shell" vs "```console"
"shell" should be used for something that you can copy-paste the entirety
of in to your terminal.

"console" should be used for things you'd see in your terminal, where not
100% of it is from you typing it--the "$ " prompt, or command output.

Sometimes something said "shell", but then had a leading "$"; if there's
no sample output after the command, just remove the "$" prompt.  If there
is output, change it to "console".

Sometimes something said "console", and had a command and output but no
"$" prompt; add the prompt.

Sometimes something said "console", and had output, but no command.  In
the cases where it's easy and readable, add the command that would
generate that output.

Sometimes something said "bash", "sh", or "shell script"; change that to
"shell".

Sometimes something said "commandline"; change that to "console".